+++ src/XpSearch.Widgets/Resources/WidgetResources.cs
@@ -30,6 +30,28 @@ internal static class WidgetResources
 
     internal static string Hint_SortOptions => Get(nameof(Hint_SortOptions));
 
+    internal static string Preview_Badge => Get(nameof(Preview_Badge));
+
+    internal static string Preview_Note_Generic => Get(nameof(Preview_Note_Generic));
+
+    internal static string Preview_Note_Attribute => Get(nameof(Preview_Note_Attribute));
+
+    internal static string Preview_Note_Results => Get(nameof(Preview_Note_Results));
+
+    internal static string Preview_Note_Suggestions => Get(nameof(Preview_Note_Suggestions));
+
+    internal static string Preview_Unset => Get(nameof(Preview_Unset));
+
+    internal static string Preview_SearchPlaceholder => Get(nameof(Preview_SearchPlaceholder));
+
+    internal static string Preview_ShowMore => Get(nameof(Preview_ShowMore));
+
+    internal static string Preview_LoadMore => Get(nameof(Preview_LoadMore));
+
+    internal static string Preview_From => Get(nameof(Preview_From));
+
+    internal static string Preview_To => Get(nameof(Preview_To));
+
     private static string Get(string name) =>
         Manager.GetString(name, CultureInfo.CurrentUICulture) ?? name;
 }